Introduced URL class & small fixes
The new URL class has replaced the Request & URI class. This will
improve readabilty and makes it easier create new URLs in modules.

This update also includes:
- define moved from config.php to index.php. This way you won't try to
redefine if you load the config a 2nd time.
- mysql fetch and fetchAll return a bool when failed, so you don't have
to use empty() to check if the query was succesfull.
- Namespace had a syntax error in the admin, install & uninstall
classes of every module (/ was used instead of \)
- Updated the admin module.